# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IPL Data Platform API",
    "version": "1.0.0",
    "description": "Ball-by-ball IPL data, modelled at one row per delivery.\n\nEvery figure this API returns is **derived from deliveries**, not copied from\nthe source feed. The vendor-supplied scorecards and league table are stored\nseparately and used only to assert that the derivation is correct — 23\ndata-quality checks run after every ingest, including one that requires the\npoints table to equal the published standings exactly, net run rate included.\n\n**Conventions**\n\n- Errors are RFC 9457 `application/problem+json`, always, with a `traceId`.\n- List endpoints use **keyset pagination**. Treat `page.nextCursor` as opaque.\n- Rates that are undefined (a strike rate off no balls) are `null`, never\n  `Infinity` or omitted.\n- Aggregates carry a strong `ETag`; send `If-None-Match` to get a 304.",
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "internalToken": {
        "type": "apiKey",
        "name": "x-internal-token",
        "in": "header",
        "description": "Shared secret guarding operational endpoints."
      }
    },
...
  },
  "paths": {
...
    "/internal/refresh-marts": {
      "post": {
        "summary": "Refresh materialised views",
        "tags": [
          "internal"
        ],
        "description": "Refreshes every mart concurrently and bumps the cache version, invalidating all cached aggregates atomically. Guarded by `x-internal-token`.",
        "security": [
          {
            "internalToken": []
          }
        ],
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "accepted"
                      ]
                    },
                    "marts": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "status",
                    "marts"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri",
                      "description": "URI identifying the problem type"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explanation specific to this occurrence"
                    },
                    "instance": {
                      "type": "string",
                      "description": "URI of the specific occurrence"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Correlates with logs and traces"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      },
                      "description": "Field-level validation failures"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
}